RDK-6313 : Ensure mLastPos is updated if throttle is disabled
Fixing bug where mLastPos was not updated if throttle was disabled. This
would cause incorrect erasing of buffers in clearToLastRead if called
when throttle was disabled. This scenario happend if channel was changed
during trick play and could mean that playback would stall which
in particular would be an issue for fast forward.

Change-Id: I6491d310eaf97a2ad887b1043783376eeb2c1075